+++ packages/terra-icon/CHANGELOG.md
@@ -16,6 +16,9 @@ Unreleased
 * Consume one-cerner-style-icons v1.26
 * Added documentation on required peerDependencies
 
+### Changed
+* Update tests to use `Terra.describeViewports` for setting viewports and use `Terra.it.validatesElement`
+
 3.13.0  - (July 11, 2019)
 ------------------
 ### Changed